+++ prism-agent/service/server/src/main/scala/io/iohk/atala/agent/server/Modules.scala
@@ -47,6 +47,7 @@ import zhttp.service.Server
 
 import java.util.concurrent.Executors
 import io.iohk.atala.mercury.*
+import io.iohk.atala.mercury.DidOps._
 import io.iohk.atala.mercury.model.*
 import io.iohk.atala.mercury.model.error.*
 import io.iohk.atala.mercury.protocol.issuecredential.*
@@ -92,12 +93,13 @@ import io.iohk.atala.agent.walletapi.storage.DIDSecretStorage
 import io.iohk.atala.pollux.vc.jwt.DidResolver as JwtDidResolver
 import io.iohk.atala.castor.core.model.error.DIDOperationError.TooManyDidServiceAccess
 import io.iohk.atala.pollux.vc.jwt.PrismDidResolver
+import io.iohk.atala.mercury.DidAgent
 
 object Modules {
 
   def app(port: Int): RIO[
-    DidComm & ManagedDIDService & AppConfig & DIDRegistrarApi & IssueCredentialsProtocolApi & ConnectionsManagementApi &
-      DIDApi & PresentProofApi & ActorSystem[Nothing],
+    DidOps & DidAgent & ManagedDIDService & AppConfig & DIDRegistrarApi & IssueCredentialsProtocolApi &
+      ConnectionsManagementApi & DIDApi & PresentProofApi & ActorSystem[Nothing],
     Unit
   ] = {
     val httpServerApp = HttpRoutes.routes.flatMap(HttpServer.start(port, _))
@@ -122,7 +124,7 @@ object Modules {
   def didCommServiceEndpoint(port: Int) = {
     val header = "content-type" -> MediaTypes.contentTypeEncrypted
     val app: HttpApp[
-      DidComm with CredentialService with PresentationService with ConnectionService & ManagedDIDService,
+      DidOps & DidAgent & CredentialService & PresentationService & ConnectionService & ManagedDIDService,
       Throwable
     ] =
       Http.collectZIO[Request] {
@@ -131,8 +133,8 @@ object Modules {
         // case Method.POST -> !! / "did-comm-v2" =>
         case Method.GET -> !! / "did" =>
           for {
-            didCommService <- ZIO.service[DidComm]
-            str = didCommService.myDid.value
+            didCommService <- ZIO.service[DidAgent]
+            str = didCommService.id.value
           } yield (Response.text(str))
 
         case req @ Method.POST -> !!
@@ -155,7 +157,7 @@ object Modules {
   }
 
   val didCommExchangesJob: RIO[
-    DIDResolver & JwtDidResolver & HttpClient & CredentialService & ManagedDIDService & DIDSecretStorage,
+    DidOps & DIDResolver & JwtDidResolver & HttpClient & CredentialService & ManagedDIDService & DIDSecretStorage,
     Unit
   ] =
     BackgroundJobs.didCommExchanges
@@ -163,14 +165,14 @@ object Modules {
       .unit
 
   val presentProofExchangeJob: RIO[
-    DIDResolver & JwtDidResolver & HttpClient & PresentationService & ManagedDIDService & DIDSecretStorage,
+    DidOps & DIDResolver & JwtDidResolver & HttpClient & PresentationService & ManagedDIDService & DIDSecretStorage,
     Unit
   ] =
     BackgroundJobs.presentProofExchanges
       .repeat(Schedule.spaced(10.seconds))
       .unit
 
-  val connectDidCommExchangesJob: RIO[DIDResolver & HttpClient & ConnectionService & ManagedDIDService, Unit] =
+  val connectDidCommExchangesJob: RIO[DidOps & DIDResolver & HttpClient & ConnectionService & ManagedDIDService, Unit] =
     ConnectBackgroundJobs.didCommExchanges
       .repeat(Schedule.spaced(10.seconds))
       .unit
@@ -192,7 +194,7 @@ object Modules {
 
   private[this] def unpackMessage(
       jsonString: String
-  ): ZIO[ManagedDIDService, ParseResponse | DIDSecretStorageError, Message] = {
+  ): ZIO[DidOps & ManagedDIDService, ParseResponse | DIDSecretStorageError, Message] = {
     // Needed for implicit conversion from didcommx UnpackResuilt to mercury UnpackMessage
     import io.iohk.atala.mercury.model.given
     for {
@@ -200,20 +202,19 @@ object Modules {
       _ <- ZIO.logInfo(s"Extracted recipient Did => $recipientDid")
       managedDIDService <- ZIO.service[ManagedDIDService]
       peerDID <- managedDIDService.getPeerDID(DidId(recipientDid))
-      msg: UnpackMessage = new DIDComm(UniversalDidResolver, peerDID.getSecretResolverInMemory).unpack(
-        new UnpackParams.Builder(jsonString).build()
-      )
+      agent = AgentPeerService.makeLayer(peerDID)
+      msg <- unpack(jsonString).provideSomeLayer(agent)
     } yield msg.message
   }
 
   def webServerProgram(
       jsonString: String
   ): ZIO[
-    CredentialService with PresentationService with ConnectionService & ManagedDIDService,
+    DidOps & CredentialService & PresentationService & ConnectionService & ManagedDIDService,
     MercuryThrowable | DIDSecretStorageError,
     Unit
   ] = {
-    import io.iohk.atala.mercury.DidComm.*
+
     ZIO.logAnnotate("request-id", java.util.UUID.randomUUID.toString()) {
       for {
         _ <- ZIO.logInfo("Received new message")
@@ -378,7 +379,7 @@ object Modules {
     }
   }
 
-  val publishCredentialsToDltJob: RIO[DidComm & CredentialService, Unit] = {
+  val publishCredentialsToDltJob: RIO[CredentialService, Unit] = {
     val effect = BackgroundJobs.publishCredentialsToDlt
     (effect repeat Schedule.spaced(1.seconds)).unit
   }
@@ -421,13 +422,13 @@ object AppModule {
     (didOpValidatorLayer ++ didServiceLayer ++ secretStorageLayer ++ nonSecretStorageLayer) >>> ManagedDIDService.layer
   }
 
-  val credentialServiceLayer: RLayer[DidComm, CredentialService] =
+  val credentialServiceLayer: RLayer[DidOps & DidAgent, CredentialService] =
     (GrpcModule.layers ++ RepoModule.credentialRepoLayer) >>> CredentialServiceImpl.layer
 
   def presentationServiceLayer =
     (RepoModule.presentationRepoLayer ++ RepoModule.credentialRepoLayer) >>> PresentationServiceImpl.layer
 
-  val connectionServiceLayer: RLayer[DidComm, ConnectionService] =
+  val connectionServiceLayer: RLayer[DidOps & DidAgent, ConnectionService] =
     (GrpcModule.layers ++ RepoModule.connectionRepoLayer) >>> ConnectionServiceImpl.layer
 
 }
@@ -481,21 +482,22 @@ object HttpModule {
   }
 
   val issueCredentialsProtocolApiLayer
-      : RLayer[DidComm & ManagedDIDService & ConnectionService & AppConfig, IssueCredentialsProtocolApi] = {
+      : RLayer[DidOps & DidAgent & ManagedDIDService & ConnectionService & AppConfig, IssueCredentialsProtocolApi] = {
     val serviceLayer = AppModule.credentialServiceLayer
     val apiServiceLayer = serviceLayer >>> IssueCredentialsProtocolApiServiceImpl.layer
     val apiMarshallerLayer = IssueCredentialsProtocolApiMarshallerImpl.layer
     (apiServiceLayer ++ apiMarshallerLayer) >>> ZLayer.fromFunction(new IssueCredentialsProtocolApi(_, _))
   }
 
-  val presentProofProtocolApiLayer: RLayer[DidComm, PresentProofApi] = {
+  val presentProofProtocolApiLayer: RLayer[DidOps & DidAgent, PresentProofApi] = {
     val serviceLayer = AppModule.presentationServiceLayer ++ AppModule.connectionServiceLayer // ++ didCommServiceLayer
     val apiServiceLayer = serviceLayer >>> PresentProofApiServiceImpl.layer
     val apiMarshallerLayer = PresentProofApiMarshallerImpl.layer
     (apiServiceLayer ++ apiMarshallerLayer) >>> ZLayer.fromFunction(new PresentProofApi(_, _))
   }
 
-  val connectionsManagementApiLayer: RLayer[DidComm & ManagedDIDService & AppConfig, ConnectionsManagementApi] = {
+  val connectionsManagementApiLayer
+      : RLayer[DidOps & DidAgent & ManagedDIDService & AppConfig, ConnectionsManagementApi] = {
     val serviceLayer = AppModule.connectionServiceLayer
     val apiServiceLayer = serviceLayer >>> ConnectionsManagementApiServiceImpl.layer
     val apiMarshallerLayer = ConnectionsManagementApiMarshallerImpl.layer